Refactor test directories
