+++ src/tsconfig.json
@@ -3,6 +3,5 @@
     "compilerOptions": {
         "rootDir": "./",
         "outDir": "../dist"
-    },
-    "includes": ["./"]
+    }
 }
