+++ tsconfig.json
@@ -9,8 +9,7 @@
         "target": "ES2020",
         "sourceMap": true,
         "incremental": true,
-        "moduleResolution": "node",
-        "baseUrl": "./"
+        "moduleResolution": "node"
     },
-    "exclude": ["node_modules", "dist"]
+    "exclude": ["node_modules", "dist", "**/*.spec.ts"]
 }
